refactor(scripts): merge x402 and MPP test scripts into unified test-payments-prod
